Create StoreModule
